Use PreferenceFragment rather than PreferenceActivity.
