+++ PuffinFeeder/src/main/java/edu/alaska/gina/feeder/puffinfeeder/PreferencesActivity.java
@@ -1,17 +1,15 @@
 package edu.alaska.gina.feeder.puffinfeeder;
 
-import android.content.SharedPreferences;
+import android.R;
+import android.app.Activity;
 import android.os.Bundle;
-import android.preference.ListPreference;
-import android.preference.Preference;
-import android.preference.PreferenceActivity;
 import android.view.MenuItem;
 
 /**
  * Activity that starts settings fragment.
  * Created by bobby on 7/25/13.
  */
-public class PreferencesActivity extends PreferenceActivity implements SharedPreferences.OnSharedPreferenceChangeListener {
+public class PreferencesActivity extends Activity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -20,28 +18,7 @@ public class PreferencesActivity extends PreferenceActivity implements SharedPre
         getActionBar().setHomeButtonEnabled(true);
         getActionBar().setTitle("Settings");
 
-        addPreferencesFromResource(R.xml.preferences);
-    }
-
-    @Override
-    protected void onResume() {
-        super.onResume();
-        getPreferenceScreen().getSharedPreferences().registerOnSharedPreferenceChangeListener(this);
-    }
-
-    @Override
-    protected void onPause() {
-        super.onPause();
-        getPreferenceScreen().getSharedPreferences().unregisterOnSharedPreferenceChangeListener(this);
-    }
-
-    @Override
-    public void onSharedPreferenceChanged(SharedPreferences sharedPreferences, String s) {
-        Preference p = findPreference(s);
-        if (p instanceof ListPreference) {
-            ListPreference l = (ListPreference) p;
-            p.setSummary(l.getEntry());
-        }
+        getFragmentManager().beginTransaction().replace(R.id.content, new SettingsFragment()).commit();
     }
 
     @Override
