+++ PuffinFeeder/src/main/java/edu/alaska/gina/feeder/puffinfeeder/SettingsFragment.java
@@ -1,0 +1,40 @@
+package edu.alaska.gina.feeder.puffinfeeder;
+
+import android.content.SharedPreferences;
+import android.os.Bundle;
+import android.preference.ListPreference;
+import android.preference.Preference;
+import android.preference.PreferenceFragment;
+
+/**
+ * Fragment that allows changes to be made to the app's global settings.
+ * Created by bobby on 8/9/13.
+ */
+public class SettingsFragment extends PreferenceFragment implements SharedPreferences.OnSharedPreferenceChangeListener {
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        addPreferencesFromResource(R.xml.preferences);
+    }
+
+    @Override
+    public void onSharedPreferenceChanged(SharedPreferences sharedPreferences, String s) {
+        Preference p = findPreference(s);
+        if (p instanceof ListPreference) {
+            ListPreference l = (ListPreference) p;
+            p.setSummary(l.getEntry());
+        }
+    }
+
+    @Override
+    public void onPause() {
+        super.onPause();
+        getPreferenceScreen().getSharedPreferences().unregisterOnSharedPreferenceChangeListener(this);
+    }
+
+    @Override
+    public void onResume() {
+        super.onResume();
+        getPreferenceScreen().getSharedPreferences().registerOnSharedPreferenceChangeListener(this);
+    }
+}
